WIP - Add Poloniex Reader
Trying to think ahead that I need
to have Poloniex and Bittrex streams
work well together, and possibly
other services, as well, I decided
to create these "readers" that will
read a stream and emit individual pieces
of data for the consumer to deal with.

In other words, I expect that a component
like "dashboard" perhaps would create these
readers and handle the smaller pieces of
data as they come in from the stream.

I think this is better than having the
stream reading and data rendering all
done in the same place and at once.