WIP: Update Poloniex Reader
- add constants to better identify
transaction types - bid, ask, etc.
in the code
- add limit to number of orderbook
records to parse from the stream
- default limit to 20
- add restart() to allow refreshing
of the orderbook after a trade
since manually calculating updated
bids/asks would be unreliable
- add option to modify the command
sent to open the websocket stream
- add more generic 'trade' event
because it's easy, and it may be
useful
- add better error checking / event
to updatePriceToTotal

Notes:
Next I'll be going over my use of
.toFixed(8), etc. I'd like to
work with floats for all calcs, etc.
and only use toFixed when printing
to the UI. Right now things are
kind of all over the place.